requirements.txt
torch
torchvision
scikit-image
scikit-learn
pyyaml
matplotlib
h5py
pycocotools
flask
flask_cors
flask_restful